+++ crates/rpc-types/src/transaction.rs
@@ -1,6 +1,6 @@
 //! Optimism specific types related to transactions.
 
-use alloy_consensus::Transaction as ConsensusTransaction;
+use alloy_consensus::Transaction as _;
 use alloy_eips::{eip2930::AccessList, eip7702::SignedAuthorization};
 use alloy_primitives::{Address, BlockHash, Bytes, ChainId, TxKind, B256, U256};
 use alloy_serde::OtherFields;
@@ -22,11 +22,14 @@ pub struct Transaction {
     #[deref_mut]
     pub inner: alloy_rpc_types_eth::Transaction<OpTxEnvelope>,
 
+    /// Nonce for deposit transactions. Only present in RPC responses.
+    pub deposit_nonce: Option<u64>,
+
     /// Deposit receipt version for deposit transactions post-canyon
     pub deposit_receipt_version: Option<u64>,
 }
 
-impl ConsensusTransaction for Transaction {
+impl alloy_consensus::Transaction for Transaction {
     fn chain_id(&self) -> Option<ChainId> {
         self.inner.chain_id()
     }
@@ -59,12 +62,20 @@ impl ConsensusTransaction for Transaction {
         self.inner.priority_fee_or_price()
     }
 
-    fn to(&self) -> Option<Address> {
-        self.inner.to()
+    fn effective_gas_price(&self, base_fee: Option<u64>) -> u128 {
+        self.inner.effective_gas_price(base_fee)
+    }
+
+    fn is_dynamic_fee(&self) -> bool {
+        self.inner.is_dynamic_fee()
     }
 
     fn kind(&self) -> TxKind {
         self.inner.kind()
+    }
+
+    fn to(&self) -> Option<Address> {
+        self.inner.to()
     }
 
     fn value(&self) -> U256 {
@@ -109,12 +120,12 @@ impl alloy_network_primitives::TransactionResponse for Transaction {
         self.inner.transaction_index()
     }
 
-    fn from(&self) -> alloy_primitives::Address {
+    fn from(&self) -> Address {
         self.inner.from()
     }
 
-    fn to(&self) -> Option<alloy_primitives::Address> {
-        ConsensusTransaction::to(&self.inner)
+    fn to(&self) -> Option<Address> {
+        alloy_consensus::Transaction::to(&self.inner)
     }
 }
 
@@ -155,15 +166,31 @@ mod tx_serde {
     //!
     //! This is needed because we might need to deserialize the `from` field into both
     //! [`alloy_rpc_types_eth::Transaction::from`] and [`op_alloy_consensus::TxDeposit::from`].
+    //!
+    //! Additionaly, we need similar logic for the `gasPrice` field
     use super::*;
     use serde::de::Error;
 
     /// Helper struct which will be flattened into the transaction and will only contain `from`
     /// field if inner [`OpTxEnvelope`] did not consume it.
     #[derive(Serialize, Deserialize)]
-    struct MaybeFrom {
+    struct OptionalFields {
         #[serde(default, skip_serializing_if = "Option::is_none")]
         from: Option<Address>,
+        #[serde(
+            default,
+            rename = "gasPrice",
+            skip_serializing_if = "Option::is_none",
+            with = "alloy_serde::quantity::opt"
+        )]
+        effective_gas_price: Option<u128>,
+        #[serde(
+            default,
+            rename = "nonce",
+            skip_serializing_if = "Option::is_none",
+            with = "alloy_serde::quantity::opt"
+        )]
+        deposit_nonce: Option<u64>,
     }
 
     #[derive(Serialize, Deserialize)]
@@ -171,19 +198,11 @@ mod tx_serde {
     pub(crate) struct TransactionSerdeHelper {
         #[serde(flatten)]
         inner: OpTxEnvelope,
-        #[serde(default, skip_serializing_if = "Option::is_none")]
+        #[serde(default)]
         block_hash: Option<BlockHash>,
-        #[serde(
-            default,
-            skip_serializing_if = "Option::is_none",
-            with = "alloy_serde::quantity::opt"
-        )]
+        #[serde(default, with = "alloy_serde::quantity::opt")]
         block_number: Option<u64>,
-        #[serde(
-            default,
-            skip_serializing_if = "Option::is_none",
-            with = "alloy_serde::quantity::opt"
-        )]
+        #[serde(default, with = "alloy_serde::quantity::opt")]
         transaction_index: Option<u64>,
         #[serde(
             default,
@@ -193,7 +212,7 @@ mod tx_serde {
         deposit_receipt_version: Option<u64>,
 
         #[serde(flatten)]
-        from: MaybeFrom,
+        other: OptionalFields,
     }
 
     impl From<Transaction> for TransactionSerdeHelper {
@@ -205,13 +224,18 @@ mod tx_serde {
                         block_hash,
                         block_number,
                         transaction_index,
+                        effective_gas_price,
                         from,
                     },
                 deposit_receipt_version,
+                deposit_nonce,
             } = value;
 
-            // if inner transaction is deposit, then don't serialize `from` directly
+            // if inner transaction is a deposit, then don't serialize `from` directly
             let from = if matches!(inner, OpTxEnvelope::Deposit(_)) { None } else { Some(from) };
+
+            // if inner transaction has its own `gasPrice` don't serialize it in this struct.
+            let effective_gas_price = effective_gas_price.filter(|_| inner.gas_price().is_none());
 
             Self {
                 inner,
@@ -219,7 +243,7 @@ mod tx_serde {
                 block_number,
                 transaction_index,
                 deposit_receipt_version,
-                from: MaybeFrom { from },
+                other: OptionalFields { from, effective_gas_price, deposit_nonce },
             }
         }
     }
@@ -234,18 +258,23 @@ mod tx_serde {
                 block_number,
                 transaction_index,
                 deposit_receipt_version,
-                from,
+                other,
             } = value;
 
             // Try to get `from` field from inner envelope or from `MaybeFrom`, otherwise return
             // error
-            let from = if let Some(from) = from.from {
+            let from = if let Some(from) = other.from {
                 from
             } else if let OpTxEnvelope::Deposit(tx) = &inner {
                 tx.from
             } else {
                 return Err(serde_json::Error::custom("missing `from` field"));
             };
+
+            // Only serialize deposit_nonce if inner transaction is deposit to avoid duplicated keys
+            let deposit_nonce = other.deposit_nonce.filter(|_| inner.is_deposit());
+
+            let effective_gas_price = other.effective_gas_price.or(inner.gas_price());
 
             Ok(Self {
                 inner: alloy_rpc_types_eth::Transaction {
@@ -254,8 +283,10 @@ mod tx_serde {
                     block_number,
                     transaction_index,
                     from,
+                    effective_gas_price,
                 },
                 deposit_receipt_version,
+                deposit_nonce,
             })
         }
     }
@@ -269,7 +300,7 @@ mod tests {
     fn can_deserialize_deposit() {
         // cast rpc eth_getTransactionByHash
         // 0xbc9329afac05556497441e2b3ee4c5d4da7ca0b2a4c212c212d0739e94a24df9 --rpc-url optimism
-        let rpc_tx = r#"{"blockHash":"0x9d86bb313ebeedf4f9f82bf8a19b426be656a365648a7c089b618771311db9f9","blockNumber":"0x798ad0b","hash":"0xbc9329afac05556497441e2b3ee4c5d4da7ca0b2a4c212c212d0739e94a24df9","transactionIndex":"0x0","type":"0x7e","nonce":"0x152ea95","input":"0x440a5e200000146b000f79c50000000000000003000000006725333f000000000141e287000000000000000000000000000000000000000000000000000000012439ee7e0000000000000000000000000000000000000000000000000000000063f363e973e96e7145ff001c81b9562cba7b6104eeb12a2bc4ab9f07c27d45cd81a986620000000000000000000000006887246668a3b87f54deb3b94ba47a6f63f32985","mint":"0x0","sourceHash":"0x04e9a69416471ead93b02f0c279ab11ca0b635db5c1726a56faf22623bafde52","r":"0x0","s":"0x0","v":"0x0","gas":"0xf4240","from":"0xdeaddeaddeaddeaddeaddeaddeaddeaddead0001","to":"0x4200000000000000000000000000000000000015","depositReceiptVersion":"0x1","value":"0x0","gasPrice":"0x0"}"#;
+        let rpc_tx = r#"{"blockHash":"0x9d86bb313ebeedf4f9f82bf8a19b426be656a365648a7c089b618771311db9f9","blockNumber":"0x798ad0b","hash":"0xbc9329afac05556497441e2b3ee4c5d4da7ca0b2a4c212c212d0739e94a24df9","transactionIndex":"0x0","type":"0x7e","nonce":"0x152ea95","input":"0x440a5e200000146b000f79c50000000000000003000000006725333f000000000141e287000000000000000000000000000000000000000000000000000000012439ee7e0000000000000000000000000000000000000000000000000000000063f363e973e96e7145ff001c81b9562cba7b6104eeb12a2bc4ab9f07c27d45cd81a986620000000000000000000000006887246668a3b87f54deb3b94ba47a6f63f32985","mint":"0x0","sourceHash":"0x04e9a69416471ead93b02f0c279ab11ca0b635db5c1726a56faf22623bafde52","r":"0x0","s":"0x0","v":"0x0","yParity":"0x0","gas":"0xf4240","from":"0xdeaddeaddeaddeaddeaddeaddeaddeaddead0001","to":"0x4200000000000000000000000000000000000015","depositReceiptVersion":"0x1","value":"0x0","gasPrice":"0x0"}"#;
 
         let tx = serde_json::from_str::<Transaction>(rpc_tx).unwrap();
 
@@ -277,5 +308,11 @@ mod tests {
             panic!("Expected deposit transaction");
         };
         assert_eq!(tx.from, inner.from);
+        assert_eq!(tx.deposit_nonce, Some(22211221));
+        assert_eq!(tx.inner.effective_gas_price, Some(0));
+
+        let deserialized = serde_json::to_value(&tx).unwrap();
+        let expected = serde_json::from_str::<serde_json::Value>(rpc_tx).unwrap();
+        similar_asserts::assert_eq!(deserialized, expected);
     }
 }