+++ CHANGELOG.md
@@ -5,6 +5,50 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.5](https://github.com/alloy-rs/op-alloy
+/releases/tag/v0.6.5) - 2024-11-12
+
+### Dependencies
+
+- Bump alloy 064 ([#249](https://github.com/alloy-rs/op-alloy/issues/249))
+
+### Features
+
+- Wrap `TxDeposit` into `Sealed` in `OpTxEnvelope` ([#247](https://github.com/alloy-rs/op-alloy/issues/247))
+- Add nonce to RPC transaction ([#246](https://github.com/alloy-rs/op-alloy/issues/246))
+
+### Miscellaneous Tasks
+
+- Add deserde test ([#248](https://github.com/alloy-rs/op-alloy/issues/248))
+
+## [0.6.4](https://github.com/alloy-rs/op-alloy
+/releases/tag/v0.6.4) - 2024-11-12
+
+### Bug Fixes
+
+- [consensus] Add conversion for `OpTxType::Eip7702` ([#244](https://github.com/alloy-rs/op-alloy/issues/244))
+- [consensus] Fix arbitrary impl for `OpTxType` ([#242](https://github.com/alloy-rs/op-alloy/issues/242))
+
+### Miscellaneous Tasks
+
+- Release 0.6.4
+- Add is dynamic fee ([#245](https://github.com/alloy-rs/op-alloy/issues/245))
+
+## [0.6.3](https://github.com/alloy-rs/op-alloy
+/releases/tag/v0.6.3) - 2024-11-08
+
+### Dependencies
+
+- Bump Alloy Deps ([#239](https://github.com/alloy-rs/op-alloy/issues/239))
+
+### Features
+
+- Bump alloy ([#240](https://github.com/alloy-rs/op-alloy/issues/240))
+
+### Miscellaneous Tasks
+
+- Release 0.6.3
+
 ## [0.6.2](https://github.com/alloy-rs/op-alloy
 /releases/tag/v0.6.2) - 2024-11-06
 
@@ -20,6 +64,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Miscellaneous Tasks
 
+- Release 0.6.2
 - Move eip1559 impls ([#237](https://github.com/alloy-rs/op-alloy/issues/237))
 - [rpc-types] Clean up Exports ([#231](https://github.com/alloy-rs/op-alloy/issues/231))
 - [book] Consolidate Links ([#230](https://github.com/alloy-rs/op-alloy/issues/230))